Use shared Gson, inner classes for JSON objects
